Add container quota support, Disable OneFS 7.2.
This really should be a major release because it breaks bc both
by dropping support for OneFS 7.2 (which went EOSL last year), and
by adding a field to IsiQuotaReq which should never have been
exported.

I'm planning a major rewrite, so this is a relatively q&d change
to support k8s_isi_provisioner.