Disable redundant code to set directory owner
The volume code made a call to set owner and group after creating the
directory. This was pointless since it's created as the user and group
specified, and it's also a problem because the attempt to set the owner
can and will fail if the parent doesn't have an ACL allowing
std_write_owner e.g. if it's simply POSIX mofe 755.
Fixes #2.